requirements.txt
discord.py>=2.0.0
python-dotenv
supabase>=1.0.0
asyncio
requests
